+++ include/fiction/technology/sidb/generators/on_the_fly_circuit_design.hpp
@@ -25,13 +25,13 @@
 #include "fiction/technology/sidb/skeleton_bestagon_library.hpp"
 #include "fiction/technology/sidb/surface_analysis.hpp"
 #include "fiction/traits.hpp"
-#include "fiction/types.hpp"
 
 #include <mockturtle/utils/stopwatch.hpp>
 
 #include <cstdio>
 #include <optional>
 #include <stdexcept>
+#include <string>
 #include <string_view>
 #include <utility>
 
@@ -46,12 +46,11 @@ class unsuccessful_pr_error : public std::runtime_error
 {
   public:
     /**
-     * This class inherits from `std::runtime_error` and is used to signal
-     * errors related to unsuccessful placement and routing.
+     * @brief Copy the message for an unsuccessful placement and routing attempt.
      *
      * @param msg The error message describing the unsuccessful placement and routing.
      */
-    explicit unsuccessful_pr_error(const std::string_view& msg) noexcept : std::runtime_error(msg.data()) {}
+    explicit unsuccessful_pr_error(const std::string_view msg) : std::runtime_error(std::string{msg}) {}
 };
 /**
  * Exception thrown if the gate design was unsuccessful. Depending on the given gate design parameters and the defect
@@ -61,13 +60,11 @@ class unsuccessful_gate_design_error : public std::runtime_error
 {
   public:
     /**
-     * This explicit constructor initializes the base `std::runtime_error` class
-     * with the provided error message, ensuring that the exception contains
-     * detailed information about the reason for the gate design failure.
+     * @brief Copy the message for an unsuccessful gate design attempt.
      *
      * @param msg A descriptive message explaining why the gate design failed.
      */
-    explicit unsuccessful_gate_design_error(const std::string_view& msg) noexcept : std::runtime_error(msg.data()) {}
+    explicit unsuccessful_gate_design_error(const std::string_view msg) : std::runtime_error(std::string{msg}) {}
 };
 /**
  * This struct stores the parameters to design an SiDB circuit on a defective surface.